🚚 Move(shared) : timeflow.js 파일 이동
- components/ 에서 밖으로 이동 -> 엄밀히 따지면 컴포넌트는 아니기 때문
